% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsts_mcmc.R
\name{mbsts.mcmc}
\alias{mbsts.mcmc}
\title{MCMC for a given MBSTS model}
\usage{
mbsts.mcmc(
  Smodel,
  X = NULL,
  H = NULL,
  nu0.r = NULL,
  s0.r,
  nu0.eps = NULL,
  s0.eps,
  niter,
  burn,
  ping = NULL
)
}
\arguments{
\item{Smodel}{A multivariate state space model of class 'SSModel'.}

\item{X}{t x P optional matrix of predictors.}

\item{H}{P x P variance-covariance matrix of the regression coefficients. Set by default to H = (X'X)^(-1).}

\item{nu0.r}{Degrees of freedom of the Inverse-Wishart prior for each Sigma.r. Set by default to n0.r = d + 2 where d is the number of time series in the multivariate model.}

\item{s0.r}{Scale matrix of the Inverse-Wishart prior for each Sigma.r.}

\item{nu0.eps}{Degrees of freedom of the Inverse-Wishart prior for Sigma.eps. Set by default to d + 2.}

\item{s0.eps}{Scale matrix of the Inverse-Wishart prior for Sigma.eps.}

\item{niter}{Number of MCMC iteration.}

\item{burn}{Desired burn-in, set by default to 0.1 * niter.}

\item{ping}{A status message it's printed every 'ping' iteration, defaults to 0.1 * 'niter'.}
}
\value{
An object of class 'mbsts' which is a list with the following components
\describe{
  \item{eta.samples}{'niter' draws from the distribution of eta_r.}
  \item{eps.samples}{'niter' draws from the distribution of eps.}
  \item{states.samples}{draws from p(alpha_t | Y_{1:T}).}
  \item{Sigma.r}{'niter' draws from the posterior distribution of Sigma.r.}
  \item{sigma.eps}{'niter' draws from the posterior distribution of Sigma.eps.}
  \item{Z.beta}{('niter'- 'burn') x P matrix of the models selected at each iteration.}
  \item{beta}{ P x d x ('niter' - 'burn') ) array of the draws from the posterior distribution of the regression coefficient matrix.}
  \item{X}{Predictor matrix.}
  \item{y}{Matrix of observations.}
  \item{Z}{(1 x m) selection matrix of the observation equation.}
  \item{T}{(m x m) matrix of the state equation.}
  \item{R}{(1 x r) matrix selecting the state disturbances.}
  \item{niter}{Number of mcmc iterations.}
  \item{burn}{Burn-in.}
}
}
\description{
MCMC to sample from the joint posterior of model parameters in an mbsts model.
}
\examples{
## Example 1 : local level + seasonal (d = 3)
y <- cbind(seq(0.5,200,by=0.5)*0.1 + rnorm(400),
           seq(100.25,200,by=0.25)*0.05 + rnorm(400),
           rnorm(400, 5,1))
model.1 <- SSModel(y ~ SSMtrend(degree = 1, Q = matrix(NA)) + SSMseasonal(period=7, Q = matrix(NA)))
mcmc.1 <- mbsts.mcmc(model.1, s0.r = diag(3), s0.eps = diag(3), niter = 100, burn = 10)

## Example 2 : local level + seasonal + covariates (d = 2)
y <- cbind(rnorm(100), rnorm(100, 2, 3))
X <- cbind(rnorm(100, 0.5, 1) + 5, rnorm(100, 0.2, 2) - 2)
model.2 <- SSModel(y ~ SSMtrend(degree = 1, Q = matrix(NA,2,2)) + SSMseasonal(period=7, Q = matrix(NA,2,2)))
mcmc.2 <- mbsts.mcmc(model.2, X = X, s0.r = diag(2), s0.eps = diag(2), niter = 100, burn = 10)
}
