% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mbsts.R
\name{plot.CausalMBSTS}
\alias{plot.CausalMBSTS}
\title{Plotting function for object of class CausalMBSTS}
\usage{
\method{plot}{CausalMBSTS}(CausalMBSTS, int.date, type = c("impact", "forecast", "ppchecks"))
}
\arguments{
\item{CausalMBSTS}{Object of class 'CausalMBSTS'}

\item{int.date}{Date of the intervention.}

\item{type}{A character string indicating the type of plot to be produced. Possible values in 'type = c("impact","forecast","ppchecks", "inclusion.probs")'.}
}
\value{

}
\description{
Function to plot the results of the causal analysis using 'causal.mbsts'
}
\details{
"impact" plots the general causal effect at every time points in the post period.
"forecast" plots the observed time series against the predicted counterfactual.
"ppchecks" draws posterior predictive checks for the model estimated in the pre period.
"inclusion.probs" plots the regressors inclusion probabilities.
}
\examples{
plot(causal)
}
