% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CausalMBSTS.R
\name{print.CausalMBSTS}
\alias{print.CausalMBSTS}
\title{Printing method for an object of class CausalMBSTS}
\usage{
\method{print}{CausalMBSTS}(x)
}
\arguments{
\item{x}{An object of class 'CausalMBSTS'.}
}
\value{

}
\description{
The method prints the results of a causal analysis in a nice and concise output format.
}
\examples{
## Example 2
set.seed(1)
t <- seq(from = 0,to = 4*pi, length.out=300)
y <- cbind(3*sin(2*t)+rnorm(300), 2*cos(2*t) + rnorm(300))
dates <- seq.Date(from = as.Date("2015-01-01"), by = "week", length.out=300)
int.date <- as.Date("2020-02-27")
y[dates >= int.date,] <- y[dates >= int.date,]+2

# Model definition
model.2 <- model(y, components = c("trend", "cycle"), cycle.period = 75)
causal.2 <- causal.mbsts(model.2, dates = dates, int.date = int.date, s0.r = 0.01*diag(2), s0.eps = 0.1*diag(2), niter = 100, burn = 10)

# Print
print(causal.2)
}
