% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CausalMBSTS.R
\name{print.summary.CausalMBSTS}
\alias{print.summary.CausalMBSTS}
\title{Format and print the estimated causal effect(s), credible interval(s), and Bayesian p-value(s) into a clear output.}
\usage{
\method{print}{summary.CausalMBSTS}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class 'summary.CausalMBSTS', a result of a call to \code{\link{summary.CausalMBSTS}}.}

\item{digits}{Number of digits to display.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
Invisibly, \code{x}.
}
\description{
Format and print the estimated causal effect(s), credible interval(s), and Bayesian p-value(s) into a clear output.
}
