% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsts_mcmc.R
\name{mbsts.mcmc}
\alias{mbsts.mcmc}
\title{MCMC for a given MBSTS model}
\usage{
mbsts.mcmc(
  Smodel,
  X,
  H = NULL,
  nu0.k = NULL,
  s0.k,
  nu0.eps = NULL,
  s0.eps,
  niter,
  burn,
  ping = NULL
)
}
\arguments{
\item{Smodel}{A multivariate state space model of class 'SSModel'.}

\item{X}{t x P matrix of predictors.}

\item{H}{P x P variance-covariance matrix of the regression coefficients. Set by default to H = (X'X)^(-1).}

\item{nu0.k}{Degrees of freedom of the Inverse-Wishart prior for each Sigma_k. Set by default to n0.k = d + 2 where d is the number of time series in the multivariate model.}

\item{s0.k}{Scale matrix of the Inverse-Wishart prior for each Sigma_k.}

\item{nu0.eps}{Degrees of freedom of the Inverse-Wishart prior for Sigma_eps. Set by default to d + 2.}

\item{s0.eps}{Scale matrix of the Inverse-Wishart prior for Sigma.eps.}

\item{niter}{Number of MCMC iteration.}

\item{burn}{Desidered burn-in, set by default to 0.1 * niter.}

\item{ping}{A status message it's printed every 'ping' iteration, defaults to 0.1 * 'niter'.}
}
\value{
An object of class 'mbsts' which is a list with the following components
\describe{   
  \item{eta.samples}{'niter' draws from the distribution of eta_k.}
  \item{eps.samples}{'niter' draws from the distribution of eps.}
  \item{states.samples}{draws from p(alpha_t | Y_{1:T}).}
  \item{sigma.eta}{'niter' draws from the posterior distribution of Sigma_eta.}
  \item{sigma.eps}{'niter' draws from the posterior distribution of Sigma_eps.}
  \item{Z.beta}{('niter'- 'burn') x P matrix of the models selected at each iteration.}
  \item{beta}{ P x d x ('niter' - 'burn') ) array of the draws from the posterior distribution of the regression coefficient matrix.}
  \item{X}{?}
  \item{y}{?}
  \item{Z}{?}
  \item{T}{?}
  \item{R}{?}
  \item{niter}{?}
  \item{burn}{?}
}
}
\description{
MCMC to sample from the joint posterior of model parameters in an mbsts model.
For now, the case without contemporaneous predictors variables is not allowed.
}
\examples{
y <- cbind(rnorm(100), rnorm(100, 2, 3))
X <- cbind(rnorm(100, 0.5, 1) + 5, rnorm(100, 0.2, 2) - 2)
model <- SSModel(y ~ SSMtrend(degree = 1, Q = matrix(NA,2,2)) + SSMseasonal(period=4, Q = matrix(NA,2,2)))
mcmc <- mbsts.mcmc(model, X = X, s0.k = diag(2), s0.eps = diag(2), niter = 100, burn = 10)
}
