% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbsts.R
\name{predict.mbsts}
\alias{predict.mbsts}
\title{Prediction for a given multivariate Bayesian structural time series model}
\usage{
\method{predict}{mbsts}(object, steps.ahead, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class 'mbsts'.}

\item{steps.ahead}{An integer value specifying the number of steps ahead to be forecasted. If 'mbsts' contains a regression component the argument is disregarded and a prediction is made with the same length of 'newdata'.}

\item{newdata}{Optional matrix of new data. Only required when 'mbsts' contains a regression component.}
\item{...}{Arguments passed to other methods (currently unused).}
}
\value{
Returns a list with the following components
\describe{
  \item{post.pred.0}{t x d x 'niter' array of in-sample forecasts.}
  \item{post.pred.1}{S x d x 'niter' array out-of-sample forecasts, where S is the number of forecasted periods (set to the length of provided new data).}
  \item{post.pred}{(t + S) x d x 'niter' array combining in- and out-of-sample forecasts.}
}
}
\description{
Given an object of class 'mbsts' and the number of 'steps.ahead' in the future to be
forecasted, this function provides in-sample forecasts and out-of-sample forecasts,
both based on drawing from the posterior predictive distribution. If 'mbsts' contains a
regression component, then the new matrix of predictors 'newdata' must be provided.
Note that NA values are not allowed in the new regressor matrix.
}
\examples{
## Example 1 :
pred.1 <- predict(mcmc.1, steps.ahead = 10)

## Example 2 :
newdata <- cbind(rnorm(30), rt(30, 2))
pred.2 <- predict(mcmc.2, newdata)
}
